+++ src/wizards/Gen1Wiz.cpp
@@ -73,9 +73,9 @@ void Gen1Wiz::set_properties() {
 void Gen1Wiz::do_layout() {
   wxFlexGridSizer *grid_sizer_1 = new wxFlexGridSizer(5, 1, 0, 0);
   wxBoxSizer *sizer_1 = new wxBoxSizer(wxHORIZONTAL);
-  grid_sizer_1->Add(label_2, 0, wxALL | wxALIGN_CENTER_HORIZONTAL, 5);
-  grid_sizer_1->Add(text_ctrl_1, 0, wxALL | wxEXPAND, 5);
-  grid_sizer_1->Add(static_line_1, 0, wxEXPAND | wxLEFT | wxRIGHT, 2);
+  grid_sizer_1->Add(label_2, 0, static_cast<int>(wxALL) | wxALIGN_CENTER_HORIZONTAL, 5);
+  grid_sizer_1->Add(text_ctrl_1, 0, static_cast<int>(wxALL) | wxEXPAND, 5);
+  grid_sizer_1->Add(static_line_1, 0, static_cast<int>(wxEXPAND) | wxLEFT | wxRIGHT, 2);
 
   if (m_warning != NULL)
     grid_sizer_1->Add(m_warning, 0, wxALL, 5);
