+++ src/wizards/LimitWiz.cpp
@@ -85,7 +85,7 @@ void LimitWiz::do_layout() {
   grid_sizer_2->Add(text_ctrl_2, 0, wxALL, 5);
   grid_sizer_2->Add(label_4, 0, wxALIGN_RIGHT | wxALIGN_CENTER_VERTICAL | wxALL,
                     5);
-  sizer_1->Add(text_ctrl_3, 0, wxALL | wxEXPAND, 5);
+  sizer_1->Add(text_ctrl_3, 0, static_cast<int>(wxALL) | wxEXPAND, 5);
   sizer_1->Add(button_1, 0, wxALL, 5);
   grid_sizer_2->Add(sizer_1, 1, 0, 0);
   grid_sizer_2->Add(label_5, 0, wxALIGN_RIGHT | wxALIGN_CENTER_VERTICAL | wxALL,
@@ -94,7 +94,7 @@ void LimitWiz::do_layout() {
   grid_sizer_2->Add(20, 20, 0, wxALL, 5);
   grid_sizer_2->Add(checkbox_1, 9, wxALL, 5);
   grid_sizer_1->Add(grid_sizer_2, 1, wxEXPAND, 0);
-  grid_sizer_1->Add(static_line_1, 0, wxEXPAND | wxLEFT | wxRIGHT, 2);
+  grid_sizer_1->Add(static_line_1, 0, static_cast<int>(wxEXPAND) | wxLEFT | wxRIGHT, 2);
   sizer_2->Add(button_2, 0, wxALL, 5);
   sizer_2->Add(button_3, 0, wxALL, 5);
   grid_sizer_1->Add(sizer_2, 1, wxALIGN_RIGHT, 0);
